Refactor exception handling and model updates for boat and client entities; add reservation functionality with validation checks

+++ tp.boat.rental/src/main/java/com/java/tp/boat/rental/model/business/Boat.java
@@ -16,8 +16,8 @@ public class Boat {
     private BoatTypes type;
     private Integer maxCapacity;
     private Float length;
-    private Float dailyRate;
-    private Float deposit;
+    private Double dailyRate;
+    private Double deposit;
     private Boolean needsLicense;
 
     public void updateWith(Boat newBoat) {
