+++ tp.boat.rental/src/main/java/com/java/tp/boat/rental/model/request/BoatCreationRequest.java
@@ -31,10 +31,10 @@ public class BoatCreationRequest {
     private Float length;
 
     @Positive(message = "Daily rate must be a positive number")
-    private Float dailyRate;
+    private Double dailyRate;
 
     @Positive(message = "Deposit must be a positive number")
-    private Float deposit;  
+    private Double deposit;  
 
     private Boolean needsLicense = false;
 }
